Start adding ability to used size-defined (instead of null terminated) strings to rapidxml.
Added ability to specify string start and end to rapidxml instead of requiring the string to be null terminated.  This was to account for the cases where N42 files had null bytes in them (multiple vendors did this!) and you wanted content after the null byte (I didnt want to have to do a pre-run thorough of the buffer since this would considerably slow down parsing all files).  I left the string null terminated implementation in for now (which one used is determined by which rapidxml::xml_document::parse is called), but I think it can be removed (so multiple versions of similar parsing code dont have to be changed for new N42 variants) after new version is tested enough.
Loosened up XML format parsing requirements (by modifying rapidxml) quite a bit.  Now files that are truncated (I've seen a number of times with N42 files) will parse, as long as a spectrum has been included.  Allow attributes to not have a value, or attribute values to not be quoted, or whitspaces in tags where there shouldnt be (also seen from a number of detectors).  Allow closing tag to be missing; works in the sense of HTML allowed to be ommitted closing tags, but now if there is an anomolous or extra, or misspelled closing tag it will cause document to end there.  Maybe loosened up in a few other things.  I created maybe a dozen test cases to check the loosened parsing works, but not a comprehensive set (I'm sure there are cases I intended to fix that arent, and other issues caused); I also ran the parser over maybe 50k files (about 1/2 were valid spectrum files) with no crashes, but further testing is needed.
Added a .gitignore file.