Fix some glitches when zooming out with a mouse.
While zooming out, the y-axis area with could change, which the mouse position was being measured relative to the spectrum area (so starting right after y-axis line), so the change in y-axis width would cause a discontinuity of the mouse position, relative to the starting (and previous) mouse positions, causing a few different glitches.
The fix was to track the mouse position as a 4-element array, instead of a 2-element array; the first two elements are the same as before (x, and y coordinates in pixels, relative to this.vis (i.e., the spectrum area)), and the 3rd and 4th elements being the x and y coordinates relative to the this.svg element (e.g., the SVG area coordinates).  The latter coordinates are then places where comparing to previous coordinates that could have had a different y-axis width.

There is likely some more cleanup and work related to these changes, especially for touch interactions.