Improve performance on mobile devices.
The "touchcancel" event wasn't being handled, and was the cause of some jitters, or od behaviors - so did an okay job trying to make things continuous over one of these events.
Fixed up a number of other touch aspects, to hopefully be more consistent and work better.
This all started from implementing using two horizontal fingers to define an ROI and have a peak fit in it, which is now fixed.