Don't abort parsing when many "invalid" lines are present.
Propagate a warning up, though, in case the user cares.

This interferes with loading valid spectrograms containing an interval where
the radiacode device reports no photons detected. This happens when the device
is briefly exposed to a fairly intense source, eg. an airport baggage scanner.

Under these conditions, the spectrogram will contain lines looking like this:
```
...
133473254726500000	1
133473254739130000	1
133473254751890000	1
...
```

Depending on the detector model, and the source dose rates, the device has
been seen to return no events for between 3 and 15 seconds.