Allow SPE files to have some non-standard header data.
Previously SPE files had to start with a "$" character - e.g., get straight to business.  But now this is loosened up, as long as "$DATA:" and "$DATE_MEA:" is within the first 1024 bytes of the file.
Also made exception in `Measurement::gamma_channel_lower(...)` and `gamma_channel_upper(...)` slightly more descriptive.