Extract `boost::hash` into a small standalone file.
Only the relevant, and used portion of `boost::hash` was extracted.
Small changes to function signatures were made to facilitate compiling, but hash function logic was left as-is, and using this extracted code seems to give same answer as original code.

This is moving towards not requiring boost to be installed to compile the code.