Promote measurement source and descriptions to be member variables of Measurement.
Previously was storing and accessing this information in the `remarks_` fields, with prefixes of "Description:" and "Source:".
Full support of all PCF fields was the impetus for these changes, but there were a few other places these fields were effectively used as well.

It would be nice to properly support this information by altering it to conform to N42-2012 <RadItemInformation> elements.