Rename RadiaCode detector type to RadiaCodeCsI10
in preparation for adding RC-103G and RC-110, both of which do not use
10x10x10mm CsI(Tl) scintillators. I'm not sure that it makes much of a
difference right now, but when new DRFs are added to InterSpec, it's
probably worth having correct information about the scintillator.

Having a hook for different detector models will probably be useful
for estimating dead time too.