Improve ability to display and interact with non-foreground peaks.
A non-backward-compatible change is the 'roiDrag' signal emmitted (when you drag the edge of an existing ROI), had to be modified; WSignal can only have up to six arguments, so to put the spectrum type into the signal, had to reduce the left and right ROI pixel values, to just thier difference.
Still more work to do, but the gist of things is to prefer matching actions against currently highlighted peaks.